refactor: consolidate test files into single test/ directory
- Move all scattered __tests__ directories into main test/ directory
- Organize tests by functionality: processors/, config/, utils/, validation/
- Remove empty test files (0 bytes)
- Clean up duplicate and backup test files
- Maintain test coverage while improving organization
