+++ test/config/loader.test.js
@@ -1,0 +1,53 @@
+const { test } = require('node:test');
+const assert = require('node:assert/strict');
+const path = require('node:path');
+const ConfigLoader = require('../loader');
+
+test('ConfigLoader', async (t) => {
+  await t.test('loads valid config from rules.yml', async () => {
+    const loader = new ConfigLoader();
+    const repoRoot = path.resolve(__dirname, '../../..');
+    const configPath = path.join(repoRoot, 'config/rules.yml');
+    // Ensure downstream loaders use the same config resolution during this test
+    process.env.CONFIG_FILE = configPath;
+    const config = loader.load(configPath);
+
+    // Basic structure checks - no version field in new structure
+    assert.ok(config.project, 'has project section');
+    assert.ok(config.automation, 'has automation section');
+    assert.ok(config.technical, 'has technical section');
+
+    // Project settings: prefer URL-based configuration which we resolve at runtime
+    assert.equal(
+      config.project.url,
+      'https://github.com/orgs/bcgov/projects/16',
+      'correct project URL'
+    );
+
+    // Automation structure
+    assert.ok(config.automation.user_scope, 'has user_scope');
+    assert.ok(config.automation.user_scope.monitored_users, 'has monitored_users');
+    assert.ok(Array.isArray(config.automation.user_scope.monitored_users), 'monitored_users is an array');
+    assert.ok(config.automation.user_scope.monitored_users.length > 0, 'has at least one monitored user');
+    assert.ok(config.automation.user_scope.monitored_users.includes('DerekRoberts'), 'includes DerekRoberts');
+    assert.ok(config.automation.repository_scope, 'has repository_scope');
+    assert.equal(config.automation.repository_scope.organization, 'bcgov', 'correct organization');
+    assert.ok(Array.isArray(config.automation.repository_scope.repositories), 'has repository list');
+    assert.ok(config.automation.repository_scope.repositories.includes('nr-nerds'), 'includes current repo');
+
+    // After normalization through loadBoardRules, check merged rules
+    const { loadBoardRules } = require('../board-rules');
+    const normalizedConfig = loadBoardRules();
+
+    const ruleSections = Object.keys(normalizedConfig.rules);
+    for (const section of ruleSections) {
+      assert.ok(Array.isArray(normalizedConfig.rules[ section ]), `has ${section} rules`);
+      assert.ok(normalizedConfig.rules[ section ].length > 0, `${section} has rules defined`);
+    }
+
+    // Technical settings
+    assert.equal(config.technical.batch_size, 10, 'correct batch size');
+    assert.equal(config.technical.update_window_hours, 24, 'correct update window');
+    assert.ok(config.technical.optimization.skip_unchanged, 'optimization enabled');
+  });
+});
