+++ test/processors/single-rule.test.js
